+++ SVFHardwareSystem.BLL/Interfaces/ProductInventoryModel.cs
@@ -18,8 +18,8 @@ namespace SVFHardwareSystem.Services.Interfaces
         public string Name { get; set; }
         public string Unit { get; set; }
         public int Year { get; set; }
-        public int Qty { get; set; }
-        public int Quantity { get { return Qty; } }
+        public decimal Qty { get; set; }
+        public decimal Quantity { get { return Qty; } }
         public decimal UnitPrice
         {
             get => TotalAmount > 0 && Qty > 0 ? TotalAmount / Qty : 0;
