+++ SVFHardwareSystem.BLL/ServiceModels/PurchaseSaleInventoryProductModel.cs
@@ -19,20 +19,20 @@ namespace SVFHardwareSystem.Services.ServiceModels
 
         public int Year { get; set; }
 
-        public int BeginningQuantity { get; set; } = 0;
+        public decimal BeginningQuantity { get; set; } = 0;
         public decimal BeginningUnitCost { get; set; } = 0;
         public decimal BeginningAmount { get { return Math.Round(BeginningQuantity * BeginningUnitCost,2); } }
 
-        public int PurchaseQuantity { get; set; } = 0;
+        public decimal PurchaseQuantity { get; set; } = 0;
         public decimal PurchaseUnitCost { get; set; } = 0;
         public decimal PurchaseAmount { get { return Math.Round(PurchaseQuantity * PurchaseUnitCost,2); } }
 
-        public int SalesQuantity { get; set; } = 0;
+        public decimal SalesQuantity { get; set; } = 0;
         public decimal SalesUnitCost { get; set; } = 0;
         public decimal SalesAmount { get { return Math.Round(SalesQuantity * SalesUnitCost,2); } }
 
 
-        public int EndingQuantity { get { return (BeginningQuantity + PurchaseQuantity) - SalesQuantity; } }
+        public decimal EndingQuantity { get { return (BeginningQuantity + PurchaseQuantity) - SalesQuantity; } }
         public decimal EndingUnitCost { get { return EndingAmount == 0 ? 0 :Math.Round(EndingAmount /EndingQuantity,2); } }
         public decimal EndingAmount { get { return (BeginningAmount + PurchaseAmount) - SalesAmount; } }
